requirements.txt
numpy~=1.26.4
PyQt6~=6.7.0
pyqtgraph~=0.13.7
openpyxl~=3.1.2
matplotlib~=3.10.0
pillow~=10.3.0
scikit-image~=0.23.2
scipy~=1.13.0
pandas~=2.2.3
xlsxwriter~=3.2.0

torch~=2.5.1
torchvision~=0.20.1